# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QUANTEX API - AI-powered technician booking platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  customer: 'customer',
  technician: 'technician',
} as const;

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: RegisterInputRole;
  /** @nullable */
  phone?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  customer: 'customer',
  technician: 'technician',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UserUpdate {
  name?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface ServiceCategory {
  id: number;
  name: string;
  icon: string;
  description: string;
}

export interface Technician {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  skills: string[];
  rating: number;
  reviewCount: number;
  isAvailable: boolean;
  completedJobs: number;
  hourlyRate: number;
  responseTime: string;
  /** @nullable */
  distance?: number | null;
  categoryIds?: number[];
}

export interface TechnicianInput {
  bio?: string;
  skills: string[];
  hourlyRate: number;
  responseTime: string;
  categoryIds?: number[];
}

export interface TechnicianUpdate {
  bio?: string;
  skills?: string[];
  hourlyRate?: number;
  responseTime?: string;
  isAvailable?: boolean;
  categoryIds?: number[];
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  accepted: 'accepted',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: number;
  customerId: number;
  technicianId: number;
  categoryId: number;
  status: BookingStatus;
  issueDescription: string;
  /** @nullable */
  address?: string | null;
  scheduledAt: string;
  estimatedCost: number;
  /** @nullable */
  finalCost?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  technicianName?: string | null;
  /** @nullable */
  categoryName?: string | null;
}

export interface BookingInput {
  technicianId: number;
  categoryId: number;
  issueDescription: string;
  address?: string;
  scheduledAt: string;
  estimatedCost: number;
  notes?: string;
}

export type BookingStatusUpdateStatus = typeof BookingStatusUpdateStatus[keyof typeof BookingStatusUpdateStatus];


export const BookingStatusUpdateStatus = {
  accepted: 'accepted',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface BookingStatusUpdate {
  status: BookingStatusUpdateStatus;
  notes?: string;
  finalCost?: number;
}

export interface TrackingInfo {
  bookingId: number;
  status: string;
  etaMinutes: number;
  technicianLat: number;
  technicianLng: number;
  progress: number;
  lastUpdated?: string;
}

export interface Review {
  id: number;
  customerId: number;
  technicianId: number;
  bookingId: number;
  rating: number;
  comment: string;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  technicianName?: string | null;
  createdAt: string;
}

export interface ReviewInput {
  technicianId: number;
  bookingId: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment: string;
}

export interface CustomerDashboard {
  activeBookings: number;
  completedBookings: number;
  totalSpent: number;
  upcomingBookings: number;
  recentBookings: Booking[];
}

export interface TechnicianDashboard {
  pendingRequests: number;
  activeJobs: number;
  completedJobs: number;
  totalEarnings: number;
  thisMonthEarnings: number;
  rating: number;
  acceptanceRate: number;
  recentBookings: Booking[];
}

export interface IssueAnalysisInput {
  /** @minLength 10 */
  description: string;
}

export type IssueAnalysisResultUrgency = typeof IssueAnalysisResultUrgency[keyof typeof IssueAnalysisResultUrgency];


export const IssueAnalysisResultUrgency = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface IssueAnalysisResult {
  category: string;
  urgency: IssueAnalysisResultUrgency;
  troubleshootingSteps: string[];
  technicianType: string;
  summary: string;
  /** @nullable */
  suggestedCategoryId?: number | null;
}

export type ListTechniciansParams = {
categoryId?: number;
available?: boolean;
limit?: number;
};

export type ListBookingsParams = {
status?: ListBookingsStatus;
role?: ListBookingsRole;
};

export type ListBookingsStatus = typeof ListBookingsStatus[keyof typeof ListBookingsStatus];


export const ListBookingsStatus = {
  pending: 'pending',
  accepted: 'accepted',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type ListBookingsRole = typeof ListBookingsRole[keyof typeof ListBookingsRole];


export const ListBookingsRole = {
  customer: 'customer',
  technician: 'technician',
} as const;

export type ListReviewsParams = {
limit?: number;
};