# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: QUANTEX API - AI-powered technician booking platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # Users
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings for current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, in_progress, completed, cancelled]
        - name: role
          in: query
          schema:
            type: string
            enum: [customer, technician]
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
  /bookings/{id}/status:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/tracking:
    get:
      operationId: getBookingTracking
      tags: [bookings]
      summary: Get live tracking info for a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tracking info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingInfo"

  # Reviews
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List all public reviews
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # Dashboard
...
  /ai/analyze-issue:
    post:
      operationId: analyzeIssue
      tags: [ai]
      summary: AI analysis of a technical issue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueAnalysisInput"
      responses:
        "200":
          description: AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueAnalysisResult"
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [customer, technician]
        phone:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # User
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, technician]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Booking:
      type: object
      required: [id, customerId, technicianId, categoryId, status, issueDescription, scheduledAt, estimatedCost, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        technicianId:
          type: integer
        categoryId:
          type: integer
        status:
          type: string
          enum: [pending, accepted, in_progress, completed, cancelled]
        issueDescription:
          type: string
        address:
          type: ["string", "null"]
        scheduledAt:
          type: string
        estimatedCost:
          type: number
        finalCost:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        customerName:
          type: ["string", "null"]
        technicianName:
          type: ["string", "null"]
        categoryName:
          type: ["string", "null"]
    BookingInput:
      type: object
      required: [technicianId, categoryId, issueDescription, scheduledAt, estimatedCost]
      properties:
        technicianId:
          type: integer
        categoryId:
          type: integer
        issueDescription:
          type: string
        address:
          type: string
        scheduledAt:
          type: string
        estimatedCost:
          type: number
        notes:
          type: string
    BookingStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, in_progress, completed, cancelled]
        notes:
          type: string
        finalCost:
          type: number
    TrackingInfo:
      type: object
      required: [bookingId, status, etaMinutes, technicianLat, technicianLng, progress]
      properties:
        bookingId:
          type: integer
        status:
          type: string
        etaMinutes:
          type: integer
        technicianLat:
          type: number
        technicianLng:
          type: number
        progress:
          type: integer
        lastUpdated:
          type: string

    # Review
    Review:
      type: object
      required: [id, customerId, technicianId, bookingId, rating, comment, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        technicianId:
          type: integer
        bookingId:
          type: integer
        rating:
          type: integer
        comment:
          type: string
        customerName:
          type: ["string", "null"]
        technicianName:
          type: ["string", "null"]
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [technicianId, bookingId, rating, comment]
      properties:
        technicianId:
          type: integer
        bookingId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string

    # Dashboard
...
    IssueAnalysisInput:
      type: object
      required: [description]
      properties:
        description:
          type: string
          minLength: 10
    IssueAnalysisResult:
      type: object
      required: [category, urgency, troubleshootingSteps, technicianType, summary]
      properties:
        category:
          type: string
        urgency:
          type: string
          enum: [low, medium, high]
        troubleshootingSteps:
          type: array
          items:
            type: string
        technicianType:
          type: string
        summary:
          type: string
        suggestedCategoryId:
          type: ["integer", "null"]