# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QUANTEX API - AI-powered technician booking platform
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  technicianId: number;
  categoryId: number;
  issueDescription: string;
  address?: string;
  scheduledAt: string;
  estimatedCost: number;
  notes?: string;
}