# lib/api-zod/src/generated/types/issueAnalysisInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QUANTEX API - AI-powered technician booking platform
 * OpenAPI spec version: 0.1.0
 */

export interface IssueAnalysisInput {
  /** @minLength 10 */
  description: string;
}