# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QUANTEX API - AI-powered technician booking platform
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  customerId: number;
  technicianId: number;
  bookingId: number;
  rating: number;
  comment: string;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  technicianName?: string | null;
  createdAt: string;
}