# lib/api-zod/src/generated/types/trackingInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QUANTEX API - AI-powered technician booking platform
 * OpenAPI spec version: 0.1.0
 */

export interface TrackingInfo {
  bookingId: number;
  status: string;
  etaMinutes: number;
  technicianLat: number;
  technicianLng: number;
  progress: number;
  lastUpdated?: string;
}